@@ -18,13 +18,6 @@ namespace Fabric_Extension_BE_Boilerplate.Utils
         private readonly ILogger<RequestLoggingFilter> _logger;
         private readonly IHttpContextAccessor _httpContextAccessor;
 
-        private static readonly List<string> HeaderKeysToLog = new List<string>
-        {
-            HttpHeaders.RequestId,
-            HttpHeaders.ActivityId,
-            HttpHeaders.XmsClientTenantId,
-        };
-
         public RequestLoggingFilter(ILogger<RequestLoggingFilter> logger, IHttpContextAccessor httpContextAccessor)
         {
             _logger = logger;
@@ -33,8 +26,8 @@ namespace Fabric_Extension_BE_Boilerplate.Utils
 
         public async Task OnActionExecutionAsync(ActionExecutingContext context, ActionExecutionDelegate next)
         {
-            var actionName = context.ActionDescriptor?.DisplayName ?? "UnknownAction";
-            var arguments = context.ActionArguments ?? new Dictionary<string, object>();
+            var actionName = context.ActionDescriptor.DisplayName;
+            var arguments = context.ActionArguments;
 
             var timestamp = DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss.fff");
             var serializedArguments = JsonConvert.SerializeObject(arguments, Formatting.None);
@@ -50,8 +43,14 @@ namespace Fabric_Extension_BE_Boilerplate.Utils
         private static IDictionary<string, string> GetHeaders(HttpContext httpContext)
         {
             var headersToLog = new Dictionary<string, string>();
-           
-            foreach (var key in HeaderKeysToLog)
+            var headerKeys = new List<string>
+            {
+                HttpHeaders.XmsClientTenantId,
+                HttpHeaders.ActivityId,
+                HttpHeaders.RequestId
+            };
+
+            foreach (var key in headerKeys)
             {
                 if (httpContext.Request.Headers.TryGetValue(key, out var headerValue))
                 {
