@@ -1,12 +1,14 @@
 import React from "react";
-import { TableMetadata } from "src/models/LakehouseExplorerModel";
+import { TableMetadata, LakehouseExplorerTablesTreeProps } from "src/models/LakehouseExplorerModel";
 import { ArrowCircleDownSplitRegular, Table20Regular } from "@fluentui/react-icons";
 import { Tree, TreeItem, TreeItemLayout, Tooltip } from "@fluentui/react-components";
 
-export function TableTreeWithSchema(tablesInLakehouse: TableMetadata[]) {
+
+export function TableTreeWithSchema(props: LakehouseExplorerTablesTreeProps) {
+    const {allTablesInLakehouse, onSelectTableCallback} = props;
     // group the tables by schema
     const tablesInLakehouseGroupedBySchema: { [key: string]: TableMetadata[] } =
-        tablesInLakehouse.reduce((acc: { [key: string]: TableMetadata[] }, table) => {
+    allTablesInLakehouse.reduce((acc: { [key: string]: TableMetadata[] }, table) => {
             if (!acc[table.schema]) {
                 acc[table.schema] = [];
             }
@@ -17,17 +19,22 @@ export function TableTreeWithSchema(tablesInLakehouse: TableMetadata[]) {
         <>
             {tablesInLakehouseGroupedBySchema &&
                 Object.keys(tablesInLakehouseGroupedBySchema).map((schema) => (
-                    <TreeItem key={schema} itemType="branch" value="Schemas">
+                    <TreeItem id={schema} key={schema} itemType="branch">
                         <Tooltip
                         relationship="label"
                         content={schema}>
-                            <TreeItemLayout className="lvl1" iconBefore={<ArrowCircleDownSplitRegular />}>
+                            <TreeItemLayout id={schema} className="lvl1" iconBefore={<ArrowCircleDownSplitRegular />}>
                                 {schema}
                             </TreeItemLayout>
                         </Tooltip>
                         <Tree selectionMode="single" size='medium'>
                             {tablesInLakehouseGroupedBySchema[schema].map((table) => (
-                                <TreeItem key={table.name} accessKey={table.path} itemType="leaf">
+                                <TreeItem 
+                                key={table.name} 
+                                accessKey={table.path} 
+                                itemType="leaf"
+                                onClick={() => onSelectTableCallback(table)}
+                                >
                                     <Tooltip
                                     relationship='label'
                                     content={table.name}>
@@ -45,3 +52,4 @@ export function TableTreeWithSchema(tablesInLakehouse: TableMetadata[]) {
         </>
     );
 }
+
