@@ -3,8 +3,13 @@ class InternalFinalJson {
         this.artifacts = [];
         this.product = null;
         this.assets = [];
-        this.extension = devParameters;
+        this.assetsV2 = {
+            images: [],
+            locales: []
+        }
         this.tabs = [];
+        this.extension = devParameters;
     }
 }
+
 module.exports = InternalFinalJson;
