@@ -6,31 +6,33 @@ const PublicTab = require('./Public/PublicTab');
 const InternalTab = require('./Internal/InternalTab');
 
 class PublicToInternalTransformer {
-    static toInternal(publicSchema, workloadName, productName = "") {
+    static toInternal(publicSchema, workloadName, isLocalized, productName = "") {
+        const prefix = isLocalized ? `${workloadName}::` : "";
+
         if (!publicSchema) {
             return null;
         }
 
         if (publicSchema instanceof PublicProduct) {
-            return this.ProductToInternal(publicSchema, workloadName);
+            return this.ProductToInternal(publicSchema, workloadName, prefix);
         } else if (publicSchema instanceof PublicItem) {
-            return this.ItemToInternal(publicSchema, workloadName, productName);
+            return this.ItemToInternal(publicSchema, workloadName, productName , prefix);
         }
     }
 
-    static ProductToInternal(publicSchema, workloadName) {
+    static ProductToInternal(publicSchema, workloadName, prefix) {
         const internalSchema = new InternalProduct(
             `${workloadName}.${publicSchema.name}`,
-            publicSchema.displayName,
-            publicSchema.fullDisplayName,
-            publicSchema.description,
+            prefix + publicSchema.displayName,
+            prefix + publicSchema.fullDisplayName,
+            prefix + publicSchema.description,
             `${workloadName}/${publicSchema.favicon}`,
             { name: `${workloadName}/${publicSchema.icon.name}` },
-            this.CreateExperienceToInternal(publicSchema.createExperience, workloadName),
-            this.WorkspaceSettingsToInternal(publicSchema.workspaceSettings, workloadName),
-            this.ProductDetailToInternal(publicSchema.productDetail, workloadName),
+            this.CreateExperienceToInternal(publicSchema.createExperience, workloadName, prefix),
+            this.WorkspaceSettingsToInternal(publicSchema.workspaceSettings, workloadName, prefix),
+            this.ProductDetailToInternal(publicSchema.productDetail, workloadName, prefix),
             publicSchema.compatibleItemTypes,
-            this.HomePageToInternal(publicSchema.homePage, workloadName),       
+            this.HomePageToInternal(publicSchema.homePage, workloadName, prefix),
         );
 
         return internalSchema;
@@ -42,8 +44,8 @@ class PublicToInternalTransformer {
         const internalItemSchema = new InternalItem(
             itemName,
             [`${workloadName}.${productName}`],
-            publicSchema.displayName,
-            publicSchema.displayNamePlural,
+            prefix + publicSchema.displayName,
+            prefix + publicSchema.displayNamePlural,
             { ...publicSchema.editor, extension: workloadName },
             { ...publicSchema.icon, name: `${workloadName}/${publicSchema.icon.name}` },
             { ...publicSchema.activeIcon, name: `${workloadName}/${publicSchema.activeIcon.name}` },
@@ -87,13 +89,15 @@ class PublicToInternalTransformer {
         };
     }
 
-    static MenuItemToInternal(publicSchema, workloadName) {
+    static MenuItemToInternal(publicSchema, workloadName, prefix) {
         if (!publicSchema || !publicSchema.handler || !publicSchema.icon) {
             return publicSchema;
         }
 
         return {
             ...publicSchema,
+            displayName: prefix + publicSchema.displayName,
+            tooltip: prefix + publicSchema.tooltip,
             handler: this.HandlerForWorkload(publicSchema.handler, workloadName),
             icon: { ...publicSchema.icon, name: `${workloadName}/${publicSchema.icon.name}` }
         };
@@ -107,25 +111,29 @@ class PublicToInternalTransformer {
         return { ...handler, extensionName: workloadName };
     }
 
-    static HomePageToInternal(publicHomePage, workloadName) {
+    static HomePageToInternal(publicHomePage, workloadName, prefix) {
         if (!publicHomePage) {
             return null;
         }
 
         return {
-            learningMaterials: publicHomePage.learningMaterials.map(lm => this.LearningMaterialForWorkload(lm, workloadName)),
+            title: publicHomePage.title ? prefix + publicHomePage.title : null,
+            description: publicHomePage.description ? prefix + publicHomePage.description : null,
+            learningMaterials: publicHomePage.learningMaterials.map(lm => this.LearningMaterialForWorkload(lm, workloadName, prefix)),
             recommendedArtifactTypes: publicHomePage.recommendedItemTypes.map(it => `${workloadName}.${it}`),
-            newSection: this.NewSectionToInternal(publicHomePage.newSection, workloadName)
+            newSection: this.NewSectionToInternal(publicHomePage.newSection, workloadName, prefix)
         };
     }
 
-    static ProductDetailToInternal(publicProductDetail, workloadName) {
+    static ProductDetailToInternal(publicProductDetail, workloadName, prefix) {
         if (!publicProductDetail) {
             return null;
         }
 
         return {
             ...publicProductDetail,
+            slogan: prefix + publicProductDetail.slogan,
+            description: prefix + publicProductDetail.description,
             image: {
                 mediaType: 0,
                 source: `${workloadName}/${publicProductDetail.image.source}`
@@ -137,58 +145,74 @@ class PublicToInternalTransformer {
         };
     }
 
-    static LearningMaterialForWorkload(material, workloadName) {
+    static LearningMaterialForWorkload(material, workloadName, prefix) {
         if (!material) {
             return null;
         }
 
         return {
             ...material,
-            onClick: this.HandlerForWorkload(material.onClick, workloadName),
+            title: prefix + material.title,
+            description: prefix + material.description,
+            introduction: material.introduction ? prefix + material.introduction: null,
+            onClick: this.LearningMaterialHandlerForWorkload(material.onClick, workloadName),
             image: `${workloadName}/${material.image}`
         };
     }
 
-    static NewSectionToInternal(publicNewSection, workloadName) {
+    static LearningMaterialHandlerForWorkload(handler, workloadName) {
+        if (!handler) {
+            // return undefined to indicate that the handler is not set and should be removed from the manifest
+            return undefined;
+        }
+
+        return { ...handler, extensionName: workloadName };
+    }
+
+    static NewSectionToInternal(publicNewSection, workloadName, prefix) {
         if (!publicNewSection) {
             return null;
         }
 
         return {
-            customActions: publicNewSection.customActions?.map(card => this.HomeCardForWorkload(card, workloadName, false))
+            customActions: publicNewSection.customActions?.map(card => this.HomeCardForWorkload(card, workloadName, prefix))
         };
     }
 
-    static HomeCardForWorkload(card, workloadName) {
+    static HomeCardForWorkload(card, workloadName, prefix) {
         if (!card) {
             return null;
         }
 
         return {
             ...card,
+            title: prefix + card.title,
             onClick: this.HandlerForWorkload(card.onClick, workloadName),
             icon: { ...card.icon, name: `${workloadName}/${card.icon.name}` },
         };
     }
 
-    static CreateExperienceToInternal(publicCreateExperience, workloadName) {
+    static CreateExperienceToInternal(publicCreateExperience, workloadName, prefix) {
         if (!publicCreateExperience) {
             return null;
         }
 
         return {
-            description: publicCreateExperience.description,
-            cards: publicCreateExperience.cards?.map(card => this.CreateCardForWorkload(card, workloadName))
+            title: publicCreateExperience.title ? prefix + publicCreateExperience.title  : null,
+            description: prefix + publicCreateExperience.description,
+            cards: publicCreateExperience.cards?.map(card => this.CreateCardForWorkload(card, workloadName, prefix))
         };
     }
 
-    static CreateCardForWorkload(card, workloadName) {
+    static CreateCardForWorkload(card, workloadName, prefix) {
         if (!card) {
             return null;
         }
 
         return {
             ...card,
+            title: prefix + card.title,
+            description: prefix + card.description,
             artifactType: `${workloadName}.${card.itemType}`,
             onClick: this.HandlerForWorkload(card.onClick, workloadName),
             icon: { ...card.icon, name: `${workloadName}/${card.icon.name}` },
